+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # 0.2.0
 
+## Feature
+
+1. 增加 log 记录
+2. 增加 ffmpegBinPath 和 logLevel 配置项
+3. 修改配置项默认地址
+
 # 0.1.1
 
 命令行修复